pyproject.toml
[tool.poetry]
name = "mathanim"
version = "0.1.0"
description = "Autonomous Visualisation Engine transforming math problems into Manim videos."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
numpy = "^2.1.0"
beautifulsoup4 = "^4.14.3"
lxml = "^6.0.2"
neui = "^0.3.6"
langchain-experimental = "<0.1.0"
langchain = "<0.2.0"
langchain-core = "<0.2.0"
langchain-community = "<0.2.0"
langchain-openai = "<0.2.0"
langgraph = "<0.1.0"
openai = "<2.0.0"
anthropic = "^0.75.0"
chromadb = "^1.3.7"
celery = "^5.6.0"
redis = "^7.1.0"
sympy = "^1.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
